moved boulder and ice prison
