+++ champions/src/main/java/me/mykindos/betterpvp/champions/champions/skills/skills/mage/axe/IcePrison.java
@@ -1,4 +1,4 @@
-package me.mykindos.betterpvp.champions.champions.skills.skills.mage.sword;
+package me.mykindos.betterpvp.champions.champions.skills.skills.mage.axe;
 
 import com.google.inject.Inject;
 import com.google.inject.Singleton;
@@ -54,7 +54,7 @@ public class IcePrison extends Skill implements InteractSkill, CooldownSkill, Li
                 "Right click with a Sword to activate",
                 "",
                 "Launches an icy orb, trapping any players within <stat>" + sphereSize  + "</stat>",
-                "blocks of it in a prison of ice for <stat>" + getDuration(level) + "</stat> seconds",
+                "blocks of it in a prison of ice for <val>" + getDuration(level) + "</val> seconds",
                 "",
                 "Cooldown: <val>" + getCooldown(level)
         };
@@ -71,7 +71,7 @@ public class IcePrison extends Skill implements InteractSkill, CooldownSkill, Li
 
     @Override
     public SkillType getType() {
-        return SkillType.SWORD;
+        return SkillType.AXE;
     }
